Move semantic_text field mappers to inference plugin
